backend/yolo-license-plate/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
asyncio==3.4.3
backcall==0.2.0
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.5
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
executing==0.10.0
fastapi==0.79.0
fastapi-mail==1.1.4
fonttools==4.37.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
imutils==0.5.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.2
opencv-python==4.6.0.66
orjson==3.7.7
packaging==21.3
pandas==1.4.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
scipy==1.9.1
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
stack-data==0.5.0
starlette==0.19.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
onnx==1.12.0
onnx-simplifier==0.4.9
onnxruntime==1.13.1